test(orchestrator): split clarification pause-drain tests into dedicated file
