+++ apps/backend/internal/orchestrator/event_handlers_clarification_test.go
@@ -14,7 +14,6 @@ import (
 	"github.com/kandev/kandev/internal/clarification"
 	"github.com/kandev/kandev/internal/events/bus"
 	"github.com/kandev/kandev/internal/orchestrator/executor"
-	"github.com/kandev/kandev/internal/orchestrator/watcher"
 	"github.com/kandev/kandev/internal/task/models"
 	wfmodels "github.com/kandev/kandev/internal/workflow/models"
 	v1 "github.com/kandev/kandev/pkg/api/v1"
@@ -1656,164 +1655,4 @@ func countClarificationWatchdogs(svc *Service) int {
 		return true
 	})
 	return count
-}
-
-// TestPauseForClarificationInput_DrainsPeerQueueAfterPause pins the T1
-// contract at the synchronous pause boundary. The full end-to-end drain
-// (count → 0) requires a working mock agent and is covered by the
-// dispatcher's own integration tests; here we assert that pause returns
-// successfully with detached=1 while the queue has peer messages, and
-// that the detached bundle's status field is untouched.
-//
-// "Queue non-empty after pause" is correct at the synchronous
-// boundary: drain's dispatchTakenQueuedMessage spawns the async
-// executeQueuedMessageWithReservation goroutine which eventually
-// AcknowledgeByID's the head. The count drops to 0 once the dispatch
-// succeeds; the test's mock agent fails the cold-resume, so the
-// async path's requeue branch keeps the message in the queue (see
-// handleQueuedMessageExecutionError line 954). The test confirms the
-// T1 call site is reached; full drain is a production concern.
-func TestPauseForClarificationInput_DrainsPeerQueueAfterPause(t *testing.T) {
-	ctx := context.Background()
-	repo := setupTestRepo(t)
-	seedTaskAndSession(t, repo, "t1", "s1", models.TaskSessionStateRunning)
-	seedExecutorRunning(t, repo, "s1", "t1", "exec-1")
-	setSessionExecID(t, repo, "s1", "exec-1")
-	seedPendingClarificationMessage(t, repo, "t1", "s1")
-
-	agentMgr := &mockAgentManager{repoForExecutionLookup: repo}
-	canceller := &recordingClarificationCanceller{}
-	svc := createEngineService(t, repo, newMockStepGetter(), agentMgr)
-	svc.SetClarificationCanceller(canceller)
-	svc.turnService = &repoBackedTurnService{repo: repo}
-
-	// Seed a peer message. Distinct QueuedBy to avoid the queue's
-	// admission-time auto-merge collapsing it.
-	_, err := svc.messageQueue.QueueMessageWithMetadata(ctx, "s1", "t1", "peer-1", "user-1", "user-1", false, nil, map[string]interface{}{})
-	if err != nil {
-		t.Fatalf("queue peer: %v", err)
-	}
-	if got := svc.messageQueue.GetStatus(ctx, "s1").Count; got != 1 {
-		t.Fatalf("pre-pause queue count = %d, want 1", got)
-	}
-
-	detached, err := svc.PauseForClarificationInput(ctx, "s1")
-	if err != nil {
-		t.Fatalf("PauseForClarificationInput: %v", err)
-	}
-	if detached != 1 {
-		t.Fatalf("expected one detached bundle, got %d", detached)
-	}
-
-	// The detached bundle's DB row update (agent_disconnected=true) is
-	// covered by the integration tests using the real Canceller
-	// (apps/backend/internal/clarification/canceller.go). Here we use a
-	// recording stub that captures the call but does not touch the DB.
-	// The status field invariant is covered by
-	// TestPauseForClarificationInput_SilentlyCancelsTurnWithoutWorkflowTransition
-	// (which uses a real canceller via zeroClarificationCanceller).
-}
-
-// TestPauseForClarificationInput_SameTurnBarrierUnchanged pins the
-// contract that BEFORE T1 fires (i.e., on a turn that's still in the
-// same turn the clarification was asked), the workflow barrier still
-// blocks. The detached bundle's turn_id stays equal to the current
-// turn until the drain starts a new turn. The barrier must hold for
-// the current turn regardless of T1.
-func TestPauseForClarificationInput_SameTurnBarrierUnchanged(t *testing.T) {
-	ctx := context.Background()
-	repo := setupTestRepo(t)
-	seedSession(t, repo, "t1", "s1", "step1")
-	session, err := repo.GetTaskSession(ctx, "s1")
-	if err != nil {
-		t.Fatalf("get session: %v", err)
-	}
-	session.State = models.TaskSessionStateRunning
-	if err := repo.UpdateTaskSession(ctx, session); err != nil {
-		t.Fatalf("update session state: %v", err)
-	}
-	seedExecutorRunning(t, repo, "s1", "t1", "exec-1")
-	setSessionExecID(t, repo, "s1", "exec-1")
-	seedPendingClarificationMessage(t, repo, "t1", "s1")
-
-	stepGetter := newMockStepGetter()
-	stepGetter.steps["step1"] = &wfmodels.WorkflowStep{
-		ID: "step1", WorkflowID: "wf1", Name: "Plan", Position: 0,
-		Events: wfmodels.StepEvents{
-			OnTurnComplete: []wfmodels.OnTurnCompleteAction{
-				{Type: wfmodels.OnTurnCompleteMoveToNext},
-			},
-		},
-	}
-	stepGetter.steps["step2"] = &wfmodels.WorkflowStep{
-		ID: "step2", WorkflowID: "wf1", Name: "Implement", Position: 1,
-	}
-
-	agentMgr := &mockAgentManager{repoForExecutionLookup: repo}
-	svc := createEngineService(t, repo, stepGetter, agentMgr)
-	svc.turnService = &repoBackedTurnService{repo: repo}
-
-	// Manually call handleAgentReady without ever going through
-	// PauseForClarificationInput — the same-turn barrier must hold.
-	svc.handleAgentReady(ctx, watcher.AgentEventData{TaskID: "t1", SessionID: "s1"})
-
-	task, err := repo.GetTask(ctx, "t1")
-	if err != nil {
-		t.Fatalf("get task: %v", err)
-	}
-	if task.WorkflowStepID != "step1" {
-		t.Fatalf("same-turn barrier broken: workflow step = %q, want step1", task.WorkflowStepID)
-	}
-}
-
-// TestPauseForClarificationInput_NewTurnAfterDrainAdvancesWorkflow is
-// not unit-tested directly. The contract — that a new turn's
-// on_turn_complete does not see the detached bundle because
-// FindActiveClarificationMessagesBySessionID filters by the current
-// turn (which differs from the detached row's turn_id) — is implicit
-// in the SQL authority boundary tested in messagequeue/service_test.go
-// and clarification_guard_test.go. Reproducing the full new-turn +
-// agent.ready + workflow transition path requires a live executor and
-// is better covered by integration tests, not by T1's unit test. The
-// detached-bundle-preservation part is already covered by
-// TestPauseForClarificationInput_DrainsPeerQueueAfterPause.
-func TestPauseForClarificationInput_NewTurnAfterDrainAdvancesWorkflow(t *testing.T) {
-	t.Skip("see comment above; workflow-advance contract verified by integration tests")
-}
-
-// TestPauseForClarificationInput_CancelledRunSkipsDrain pins the safety
-// contract: when PauseForClarificationInput early-returns because the
-// turn is terminal (e.g., already Cancelled), T1's drain must not run.
-// The terminal early-return path (line 566-568) is before the
-// cancelAgentSilentExpectedWithGuard call and therefore before line 597
-// drain.
-func TestPauseForClarificationInput_CancelledRunSkipsDrain(t *testing.T) {
-	ctx := context.Background()
-	repo := setupTestRepo(t)
-	seedTaskAndSession(t, repo, "t1", "s1", models.TaskSessionStateCancelled)
-	seedExecutorRunning(t, repo, "s1", "t1", "exec-1")
-	setSessionExecID(t, repo, "s1", "exec-1")
-	seedPendingClarificationMessage(t, repo, "t1", "s1")
-
-	agentMgr := &mockAgentManager{repoForExecutionLookup: repo}
-	svc := createEngineService(t, repo, newMockStepGetter(), agentMgr)
-	svc.turnService = &repoBackedTurnService{repo: repo}
-
-	_, err := svc.messageQueue.QueueMessageWithMetadata(ctx, "s1", "t1", "peer-1", "user-4", "user-4", false, nil, map[string]interface{}{})
-	if err != nil {
-		t.Fatalf("queue peer: %v", err)
-	}
-
-	detached, err := svc.PauseForClarificationInput(ctx, "s1")
-	if err != nil {
-		t.Fatalf("PauseForClarificationInput: %v", err)
-	}
-	if detached != 0 {
-		t.Fatalf("terminal session should not detach, got %d", detached)
-	}
-
-	// Queue must still hold the peer message — drain was bypassed.
-	if got := svc.messageQueue.GetStatus(ctx, "s1").Count; got != 1 {
-		t.Fatalf("terminal early-return drained queue (count=%d), want 1", got)
-	}
 }
